trocando ids de Long para Integer e criando todas as entities

+++ app/src/main/java/com/tcc/zipzop/activity/ItemActivity.java
@@ -89,7 +89,7 @@ public class ItemActivity extends AppCompatActivity {
     @Override
     public boolean onContextItemSelected(@NonNull MenuItem menuItem) {
 
-        Long id = itemAdapterActivity.getId();
+        Integer id = itemAdapterActivity.getId();
         Item item = new Item();
         try {
             item = new ConsultarItemTask(dao, id).execute().get();
